refactor(wasm): move wasm_bindgen support to js folder

refactor(wasm): add wasm mod
